Add basic structure for testing conversion to & from markdown
Note: It's not quite working yet. Changing something in the BlockNote editor
correctly transfers it into the markdown textarea. If something gets lost in
the transformation (such as a nested block), the button can be used to sync this
lossy information back, causing the BlockNote to also lose it (as expected, but not desired).

However, directly editing the markdown causes the transformation to break in both directions.
This could very well be related to the Yjs merge mechanism, but I don't yet understand what's
happening.