[#68978] Change approach for dark theme status and type colors
https://community.openproject.org/wp/68978

Unfortunately I was not able to find out what was wrong with the
approach used so far. However, text colors and also the status
border colors were off (for the status colors it was not as visible
but the status text color was very clearly not what was expected).

Due to time pressure of the 17.0 release I switched the approach
so this is looking good and gets done in time. The way it is
implemented now closely resembles the colors_helper.rb in OpenProject.
Conceptually, the other approach would have been better because it
would probably be easier to memoize / cache values for re-using them.